requirements.txt
Flask
requests
python-dotenv
gunicorn
pillow
pytesseract
PyMuPDF